Refactor: merge Admin(Crawler) into crawler
